Moving high-frequency event listener in new class
